refactor: Extract shared message types to responses/types.rs
Phase 6, Step 1: Eliminate code duplication between Conversations and Responses APIs

Changes:
- Created src/web/responses/types.rs with shared types:
  - MessageContentPart enum (input message content parts)
  - MessageContent enum (text or multipart content)
  - ConversationContent enum (output message content)
  - From<MessageContent> conversion impl

- Updated src/web/responses/mod.rs to export new types module

- Removed 125 lines of duplicate code from src/web/conversations.rs:
  - Removed MessageContentPart enum definition
  - Removed MessageContent enum and impl methods
  - Removed ConversationContent enum definition
  - Removed From<MessageContent> conversion impl

- Updated imports across all modules:
  - responses/conversions.rs: Import from responses::types
  - responses/handlers.rs: Import types from responses module
  - responses/context_builder.rs: Use MessageContentConverter
  - conversations.rs: Import types from responses module

Impact:
- Zero code duplication - single source of truth for shared types
- conversations.rs: 952 → 827 lines (-125 lines, -13.1% reduction)
- 100% backward compatible - no API changes
- All external behavior unchanged
