+++ src/web/responses/mod.rs
@@ -12,6 +12,7 @@ pub mod events;
 pub mod handlers;
 pub mod storage;
 pub mod stream_processor;
+pub mod types;
 
 // Re-export commonly used types
 pub use builders::{build_usage, ContentPartBuilder, OutputItemBuilder, ResponseBuilder};
@@ -21,6 +22,7 @@ pub use errors::error_mapping;
 pub use events::{ResponseEvent, SseEventEmitter};
 pub use storage::storage_task;
 pub use stream_processor::UpstreamStreamProcessor;
+pub use types::{ConversationContent, MessageContent, MessageContentPart};
 
 // Re-export the router from handlers (legacy file)
 pub use handlers::router;
